Work on main. Needs fixations.

+++ src/main/java/bgu/spl/mics/application/TickBroadcast.java
@@ -3,17 +3,18 @@ package bgu.spl.mics.application;
 import bgu.spl.mics.Broadcast;
 
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 public class TickBroadcast implements Broadcast {
 
-    private AtomicLong currentTime;
+    private AtomicInteger currentTime;
 
-    public TickBroadcast(long CurrentTime){
-        currentTime = new AtomicLong(CurrentTime);
+    public TickBroadcast(int CurrentTime){
+        currentTime = new AtomicInteger(CurrentTime);
     }
 
-    public long getCurrentTime() {
+    public int getCurrentTime() {
         return currentTime.get();
     }
 }
